refactor: rename buff-condition-factory to alteration-condition-factory
Aligns the filename with the exported function name `buildAlterationCondition`
and the broader `alteration` vocabulary used throughout the codebase.

Closes #279
