+++ src/fight/http-api/alteration-condition-factory.ts
